Move codec/header.rs to throttled/codec.rs.
