+++ protocols/request-response/src/codec.rs
@@ -18,8 +18,6 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-pub(crate) mod header;
-
 pub use libp2p_core::ProtocolName;
 
 use async_trait::async_trait;
